+++ crates/augmented/data/augmented-midi/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.2.0
+
+* [`b0bfa7ee6`](https://github.com/yamadapc/augmented-audio/commits/b0bfa7ee6) audio-processor-traits@2.1.0 (Minor)
+* [`dcfd25994`](https://github.com/yamadapc/augmented-audio/commits/dcfd25994) Move to mainline midir :patch: (Patch)
+
 ## v1.1.0
 
 * [`ced6158a3`](https://github.com/yamadapc/augmented-audio/commits/ced6158a3) audio-processor-traits@1.1.0 (Minor)