+++ crates/augmented/dsp/convert-sample-rate/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v1.2.0
+
+* [`27535547c`](https://github.com/yamadapc/augmented-audio/commits/27535547c) audio-processor-testing-helpers@1.2.0 (Minor)
+
 ## v1.1.0
 
 * [`19cc6930f`](https://github.com/yamadapc/augmented-audio/commits/19cc6930f) audio-processor-testing-helpers@1.1.0 (Minor)